requirements.txt
Django>=5.0
psycopg2-binary>=2.9
django-widget-tweaks>=1.5.0